image pipeline has a working png

+++ src/film.jl
@@ -173,7 +173,7 @@ function merge_film_tile!(f::Film, ft::FilmTile)
     end
 end
 
-function save(film::Film, splat_scale::Float32 = 1)
+function save(film::Film, splat_scale::Float64 = 1.0)
     image = Array{Float32}(undef, size(film.pixels)..., 3)
     for y in 1:size(film.pixels)[1], x in 1:size(film.pixels)[2]
         pixel = film.pixels[y, x]
